+++ Changelog.md
@@ -16,6 +16,7 @@
   * package URL
   * package source RPM file name
   * package group name
+  * package summary
 
 ## 0.0.2